api/requirements.txt
anyio==4.5.2
attrs==25.3.0
Authlib==1.3.2
azure-core==1.32.0
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
azure-storage-blob==12.24.0
bcrypt==4.2.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
CouchDB==1.2
cryptography==44.0.0
cycler==0.12.1
exceptiongroup==1.2.2
filelock==3.16.1
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.55.0
fsspec==2024.10.0
greenlet==3.2.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mistune==3.1.3
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
networkx==3.1
ngrok==1.4.0
numpy==1.24.4
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.0.3
pillow==10.4.0
portalocker==2.10.1
psutil==6.1.0
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scipy==1.10.1
seaborn==0.13.2
shapely==2.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.39
ultralytics-thop==2.0.12
urllib3==2.2.3
Werkzeug==3.0.6
yt-dlp==2025.3.31
zipp==3.20.2
